# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/bags": {
      "get": {
        "operationId": "BagsController_list",
        "parameters": [
          {
            "name": "lat",
            "required": false,
            "in": "query",
            "description": "Kullanıcı enlemi",
            "schema": {
              "example": 40.9877,
              "type": "number"
            }
          },
          {
            "name": "lng",
            "required": false,
            "in": "query",
            "description": "Kullanıcı boylamı",
            "schema": {
              "example": 29.0277,
              "type": "number"
            }
          },
          {
            "name": "radiusKm",
            "required": false,
            "in": "query",
            "description": "Arama yarıçapı (km)",
            "schema": {
              "default": 10,
              "type": "number"
            }
          },
          {
            "name": "category",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "BAKERY",
                "MARKET",
                "CAFE",
                "RESTAURANT"
              ]
            }
          },
          {
            "name": "pickupFrom",
            "required": false,
            "in": "query",
            "description": "Teslim aralığı başlangıç filtresi",
            "schema": {
              "example": "18:00",
              "type": "string"
            }
          },
          {
            "name": "pickupTo",
            "required": false,
            "in": "query",
            "schema": {
              "example": "23:00",
              "type": "string"
            }
          },
          {
            "name": "maxPriceMinor",
            "required": false,
            "in": "query",
            "description": "Azami satış fiyatı (kuruş)",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "İşletme veya paket adında arama",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "schema": {
              "default": "relevance",
              "type": "string",
              "enum": [
                "relevance",
                "distance",
                "price",
                "rating",
                "pickup_time"
              ]
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "maximum": 50,
              "default": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Paketleri listeler (nearby ile aynı filtreler)",
        "tags": [
          "bags"
        ]
      }
    },
    "/v1/bags/{id}": {
      "get": {
        "operationId": "BagsController_detail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Paket detayı",
        "tags": [
          "bags"
        ]
      }
    },
...
    "/v1/orders": {
      "post": {
        "description": "Stok, ödeme onayı beklenmeden rezerve edilir; aksi hâlde ödeme sayfasındayken paket başkasına satılabilirdi. Ödeme tamamlanmazsa rezervasyon otomatik geri verilir.",
        "operationId": "OrdersController_create",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "Kullanıcı eylemi başına bir kez üretilmelidir (istek başına değil). Ağ tekrarında aynı anahtarla gelen istek ikinci sipariş oluşturmaz.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrderDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Sipariş oluşturur ve ödemeyi başlatır",
        "tags": [
          "orders"
        ]
      },
      "get": {
        "operationId": "OrdersController_list",
        "parameters": [
          {
            "name": "status",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Kullanıcının siparişleri",
        "tags": [
          "orders"
        ]
      }
    },
...
    "/v1/partner/stores": {
      "get": {
        "operationId": "PartnerController_stores",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Kullanıcının yönetebildiği işletmeler",
        "tags": [
          "partner"
        ]
      }
    },
    "/v1/partner/dashboard": {
      "get": {
        "description": "Bugünün cirosu, bekleyen teslimler, aktif paketler ve 7 günlük seri tek çağrıda döner.",
        "operationId": "PartnerController_dashboard",
        "parameters": [
          {
            "name": "storeId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Panel özeti",
        "tags": [
          "partner"
        ]
      }
    },
    "/v1/partner/bags": {
      "get": {
        "operationId": "PartnerController_listBags",
        "parameters": [
          {
            "name": "storeId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "İşletmenin paketleri (satılan adet dahil)",
        "tags": [
          "partner"
        ]
      },
      "post": {
        "operationId": "PartnerController_createBag",
        "parameters": [
          {
            "name": "storeId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBagDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Yeni sürpriz paket yayınlar",
        "tags": [
          "partner"
        ]
      }
    },
    "/v1/partner/bags/{id}": {
      "patch": {
        "description": "Toplam adet, satılmış adedin altına indirilemez.",
        "operationId": "PartnerController_updateBag",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBagDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Paketi günceller",
        "tags": [
          "partner"
        ]
      },
      "delete": {
        "description": "Sipariş almış paket silinmez, iptal edilir — geçmiş bozulmamalı.",
        "operationId": "PartnerController_deleteBag",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Paketi siler",
        "tags": [
          "partner"
        ]
      }
    },
    "/v1/partner/bags/{id}/publish": {
      "post": {
        "operationId": "PartnerController_publishBag",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Paketi yayına alır",
        "tags": [
          "partner"
        ]
      }
    },
    "/v1/partner/bags/{id}/pause": {
      "post": {
        "operationId": "PartnerController_pauseBag",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Paketi yayından kaldırır",
        "tags": [
          "partner"
        ]
      }
    },
    "/v1/partner/bags/{id}/toggle": {
      "post": {
        "operationId": "PartnerController_toggleBag",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ToggleBagDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Paketi yayına alır veya kaldırır",
        "tags": [
          "partner"
        ]
      }
    },
    "/v1/partner/templates": {
      "get": {
        "operationId": "PartnerController_listTemplates",
        "parameters": [
          {
            "name": "storeId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Tekrar eden paket şablonları",
        "tags": [
          "partner"
        ]
      },
      "post": {
        "description": "Zamanlanmış iş, aktif şablonlardan her gün paket üretir.",
        "operationId": "PartnerController_createTemplate",
        "parameters": [
          {
            "name": "storeId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBagTemplateDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Tekrar eden paket şablonu oluşturur",
        "tags": [
          "partner"
        ]
      }
    },
    "/v1/partner/templates/{id}/toggle": {
      "post": {
        "operationId": "PartnerController_toggleTemplate",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ToggleBagDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Şablonu etkinleştirir veya durdurur",
        "tags": [
          "partner"
        ]
      }
    },
    "/v1/partner/orders": {
      "get": {
        "description": "En yeniden eskiye sıralanır ve sayfalanır. Müşteri adı maskelenir; teslimde eşleştirme için baş harf yeterlidir.",
        "operationId": "PartnerController_listOrders",
        "parameters": [
          {
            "name": "storeId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "date",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "İşletmenin siparişleri",
        "tags": [
          "partner"
        ]
      }
    },
    "/v1/partner/orders/{id}/confirm-pickup": {
      "post": {
        "description": "Müşteri uygulamayı açamadığında (şarj bitti, internet yok) personel kodu girerek teslimi tamamlayabilir.",
        "operationId": "PartnerController_confirmPickup",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmPartnerPickupDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Teslim kodunu doğrulayarak siparişi tamamlar",
        "tags": [
          "partner"
        ]
      }
    },
    "/v1/partner/store": {
      "get": {
        "operationId": "PartnerController_store",
        "parameters": [
          {
            "name": "storeId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Mağaza profili",
        "tags": [
          "partner"
        ]
      },
      "patch": {
        "operationId": "PartnerController_updateStore",
        "parameters": [
          {
            "name": "storeId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStoreDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Mağaza profilini günceller",
        "tags": [
          "partner"
        ]
      }
    },
    "/v1/partner/payouts": {
      "get": {
        "operationId": "PartnerController_payoutHistory",
        "parameters": [
          {
            "name": "storeId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Hakediş geçmişi",
        "tags": [
          "partner"
        ]
      }
    },
    "/v1/partner/payouts/summary": {
      "get": {
        "description": "Yalnızca teslim alınmış siparişler hakedişe girer; iade edilenler dönemden düşülür.",
        "operationId": "PartnerController_payoutSummary",
        "parameters": [
          {
            "name": "storeId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "month",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Dönem hakediş özeti",
        "tags": [
          "partner"
        ]
      }
    },
    "/v1/partners/applications": {
      "post": {
        "operationId": "PartnerApplicationController_apply",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PartnerApplicationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "İşletme ön kayıt başvurusu",
        "tags": [
          "partner"
        ]
      }
    },
...
  },
  "info": {
    "title": "YePaket API",
    "description": "Gıda kurtarma pazaryeri API’si. Tüm parasal alanlar tam sayı kuruştur ve adı `_minor` ile biter. Zaman damgaları UTC ve ISO 8601’dir.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateOrderDto": {
        "type": "object",
        "properties": {
          "bagId": {
            "type": "string",
            "description": "Sipariş verilecek paket"
          },
          "quantity": {
            "type": "object",
            "minimum": 1,
            "maximum": 10,
            "default": 1
          }
        },
        "required": [
          "bagId",
          "quantity"
        ]
      },
...
      "CreateBagDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "Günün Fırın Paketi"
          },
          "category": {
            "type": "string",
            "enum": [
              "BAKERY",
              "MARKET",
              "CAFE",
              "RESTAURANT"
            ]
          },
          "description": {
            "type": "string"
          },
          "imageUrls": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "originalValueMinor": {
            "type": "number",
            "description": "Normal değer (kuruş)",
            "example": 42000
          },
          "salePriceMinor": {
            "type": "number",
            "description": "Satış fiyatı (kuruş)",
            "example": 13900
          },
          "quantity": {
            "type": "number",
            "minimum": 1,
            "maximum": 200
          },
          "pickupStartsAt": {
            "type": "string",
            "description": "Teslim başlangıcı (ISO-8601 UTC)"
          },
          "pickupEndsAt": {
            "type": "string",
            "description": "Teslim bitişi (ISO-8601 UTC)"
          }
        },
        "required": [
          "title",
          "category",
          "originalValueMinor",
          "salePriceMinor",
          "quantity",
          "pickupStartsAt",
          "pickupEndsAt"
        ]
      },
      "UpdateBagDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "quantity": {
            "type": "number",
            "description": "Yeni toplam adet"
          },
          "salePriceMinor": {
            "type": "number"
          },
          "pickupStartsAt": {
            "type": "string"
          },
          "pickupEndsAt": {
            "type": "string"
          }
        }
      },
      "ToggleBagDto": {
        "type": "object",
        "properties": {
          "published": {
            "type": "boolean"
          }
        },
        "required": [
          "published"
        ]
      },
      "CreateBagTemplateDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "category": {
            "type": "string",
            "enum": [
              "BAKERY",
              "MARKET",
              "CAFE",
              "RESTAURANT"
            ]
          },
          "description": {
            "type": "string"
          },
          "originalValueMinor": {
            "type": "number"
          },
          "salePriceMinor": {
            "type": "number"
          },
          "defaultQuantity": {
            "type": "number"
          },
          "pickupStart": {
            "type": "string",
            "example": "20:00",
            "description": "Yerel saat"
          },
          "pickupEnd": {
            "type": "string",
            "example": "20:30",
            "description": "Yerel saat"
          },
          "publishMode": {
            "type": "string",
            "enum": [
              "ONCE",
              "DAILY",
              "WEEKLY"
            ]
          },
          "weekdays": {
            "description": "WEEKLY modunda yayın günleri (1=Pazartesi … 7=Pazar)",
            "type": "array",
            "items": {
              "type": "number"
            }
          }
        },
        "required": [
          "title",
          "category",
          "originalValueMinor",
          "salePriceMinor",
          "defaultQuantity",
          "pickupStart",
          "pickupEnd",
          "publishMode"
        ]
      },
      "ConfirmPartnerPickupDto": {
        "type": "object",
        "properties": {
          "pickupCode": {
            "type": "string",
            "description": "Müşterinin gösterdiği 6 haneli teslim kodu"
          }
        },
        "required": [
          "pickupCode"
        ]
      },
      "UpdateStoreDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "addressLine": {
            "type": "string"
          },
          "district": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "latitude": {
            "type": "number"
          },
          "longitude": {
            "type": "number"
          },
          "openingTime": {
            "type": "string",
            "example": "07:30"
          },
          "closingTime": {
            "type": "string",
            "example": "21:30"
          },
          "logoUrl": {
            "type": "string"
          },
          "coverUrl": {
            "type": "string"
          }
        }
      },
      "PartnerApplicationDto": {
        "type": "object",
        "properties": {
          "businessName": {
            "type": "string"
          },
          "businessType": {
            "type": "string",
            "example": "Fırın / Pastane"
          },
          "contactName": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "district": {
            "type": "string"
          },
          "note": {
            "type": "string"
          }
        },
        "required": [
          "businessName",
          "businessType",
          "contactName",
          "phone",
          "email",
          "city",
          "district"
        ]
      },
...
    }
  }
}